Rename package: rttptool -> templtest
