+++ tests/templtest/discovery/test_Variables.py
@@ -3,8 +3,8 @@
 from pathlib import Path
 from unittest import TestCase
 
-from rttptool.discovery import Variables
-from rttptool.exception import TestDefinitionError
+from templtest.discovery import Variables
+from templtest.exception import TestDefinitionError
 
 
 class CreateVariables(TestCase):
