+++ tests/test_util/test_TempDirectory.py
@@ -10,5 +10,5 @@ class TestTempDirectory(TestCase):
     def test(self):
         with TempDirectory() as temp_dir_path:
             self.assertTrue(temp_dir_path.is_dir())
-            self.assertTrue(temp_dir_path.name.startswith("rttptool.test."))
+            self.assertTrue(temp_dir_path.name.startswith("templtest.test."))
         self.assertFalse(temp_dir_path.exists())
